Move add function into separate file
