+++ src/main/kotlin/com/manic/skeleton/Add.kt
@@ -1,0 +1,23 @@
+package com.manic.skeleton
+
+import com.xenomachina.argparser.SystemExitException
+import java.io.File
+
+fun add(addCommandArgs: AddCommandArgs) {
+    val domainName = addCommandArgs.domainName
+    val parent = File(domainName)
+
+    if (parent.exists()) {
+        throw SystemExitException("Directory ${parent.absoluteFile} is already existing.", 101)
+    }
+
+    val application = File("$domainName/$domainName-application")
+    val domain = File("$domainName/$domainName-domain")
+    val infrastructure = File("$domainName/$domainName-infrastructure")
+    arrayOf(
+        parent,
+        application,
+        domain,
+        infrastructure
+    ).forEach { it.mkdir() }
+}
